refactoring code with custom hooks
